% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_boosting.R
\name{grad_boost}
\alias{grad_boost}
\title{Fit a boosted linear model}
\usage{
grad_boost(formula, data, nu = 0.01, stop, grad.fun, loss.fun, yhat.init = 0)
}
\arguments{
\item{formula}{an object of class formula}

\item{data}{a data.frame or matrix}

\item{nu}{a numeric within the range of \link{0,1}, the learning rate}

\item{stop}{a numeric value determining the total boosting iterations}

\item{grad.fun}{a function which computes the gradient of the loss function}

\item{loss.fun}{a convex loss function which is continuously differentiable}

\item{yhat.init}{a numeric value determining the starting point}
}
\value{
\itemize{
\item theta - this is our estimator
\item u - this is the last gradient value
\item fit - our fitted values, i.e. X \%*\% theta
\item formula - the underlying formula
\item data - the underlying data
}
}
\description{
Fit a boosted linear model
}
\examples{
# Complete runthrough see: www.github.com/andrebleier/cheapml
}
