% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_rf.R
\name{reg_rf}
\alias{reg_rf}
\title{reg_rf
Fits a random forest with a continuous scaled features and target
variable (regression)}
\usage{
reg_rf(formula, n_trees, feature_frac, data)
}
\arguments{
\item{formula}{an object of class formula}

\item{n_trees}{an integer specifying the number of trees to sprout}

\item{feature_frac}{an numeric value defined between \link{0,1}
specifies the percentage of total features to be used in
each regression tree}

\item{data}{a data.frame or matrix}
}
\value{

}
\description{
reg_rf
Fits a random forest with a continuous scaled features and target
variable (regression)
}
\examples{
# Complete runthrough see: www.github.com/andrebleier/cheapml
}
