% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_tree_imp.R
\name{reg_tree_imp}
\alias{reg_tree_imp}
\title{reg_tree
Fits a simple regression tree with SSE splitting criterion. The estimator function
is the mean.}
\usage{
reg_tree_imp(formula, data, minsize)
}
\arguments{
\item{formula}{an object of class formula}

\item{data}{a data.frame or matrix}

\item{minsize}{a numeric value indicating the minimum size of observations
in a leaf}
}
\value{
\itemize{
\item tree - the tree object containing all splitting rules and observations
\item imp - returns the feature importance
\item fit - our fitted values, i.e. X \%*\% theta
\item formula - the underlying formula
\item data - the underlying data
}
}
\description{
reg_tree
Fits a simple regression tree with SSE splitting criterion. The estimator function
is the mean.
}
\examples{
# Complete runthrough see: www.github.com/andrebleier/cheapml
}
