% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_tree_imp.R
\name{reg_tree_imp}
\alias{reg_tree_imp}
\title{reg_tree
Fits a simple regression tree with SSE splitting criterion. The estimator function
is the mean.}
\usage{
reg_tree_imp(formula, data, minsize)
}
\arguments{
\item{formula}{an object of class formula}

\item{data}{a data.frame or matrix}

\item{minsize}{a numeric value indicating the minimum size of observations
in a leaf}
}
\value{
\itemize{
\item tree - the tree object containing all splitting rules and observations
\item imp - returns the feature importance
\item fit - our fitted values, i.e. X \%*\% theta
\item formula - the underlying formula
\item data - the underlying data
}
}
\description{
reg_tree
Fits a simple regression tree with SSE splitting criterion. The estimator function
is the mean.
}
\examples{
# Complete runthrough see: www.github.com/andrebleier/cheapml
# Regression Tree Runthrough ----------------------------------------------

# load data simulation tool
# library(devtools)
# devtools::install_github('andrebleier/Xy')
library(Xy)

# source the fitting function
source("reg_tree_imp.R")

# simulate data
sim <- Xy(n = 2500, # 2500 observations
         numvars = c(3, 0), # 2 linear variables / 0 nonlinear
         noisevars = 0, # no noise variables
         catvars = 0, # omit dummy variables
         stn = 5,
         intercept = FALSE
)

# simulation overview
sim

# get the formula
eq <- sim$eq

# get the formula
model_df <- sim$data

# source function
source("algorithms/reg_tree_imp.R")

# fit
mod <- reg_tree_imp(formula = eq, data = model_df, minsize = 100)

# our tree
mod$tree

# the feature importance
mod$imp

# check the feature importance of the simulation
varimp(sim)
}
